+++ src/main/java/com/put/bd/pizzeria/persistance/ClientRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClientRepository extends JpaRepository<Client, Long> {
+public interface ClientRepository extends JpaRepository<Client, Integer> {
 
     List<Client> findByFirstNameAndLastName(String firstName, String lastName);
